requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
Flask==3.0.2
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.3.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.8.3
numpy==1.26.4
openai==1.14.0
packaging==24.2
pandas==2.2.1
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
scikit-learn==1.4.1.post1
scipy==1.15.2
seaborn==0.13.1
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
